Change the flusher logic to create a thread per bucket
Previously we were creating a single thread for all instances of Cantor Events meaning that buckets A and B could not flush at the same time. We did this because it's possible to create multiple instances for the same bucket which with older logic use to accidentally trigger multiple uploads of the same bucket in one minute. This now creates a thread per bucket, but only ever a single one.